Reorganizing existing histogram test with other density utils tests
